refactor: replace template placeholders
Follow the README and TODOs to replace all template placeholders with
suitable values for the tis-trainee-notifications service.

TIS21-5056
TIS21-5055
