+++ .aws/task-definition-template.json
@@ -1,44 +1,44 @@
 {
   "containerDefinitions": [
     {
-      "name": "tis-template",
-      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-template:latest",
+      "name": "tis-trainee-notifications",
+      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-trainee-notifications:latest",
       "secrets": [
         {
           "name": "SENTRY_DSN",
-          "valueFrom": "tis-template-sentry-dsn"
+          "valueFrom": "tis-trainee-notifications-sentry-dsn"
         }
       ],
       "logConfiguration": {
         "logDriver": "awslogs",
         "options": {
-          "awslogs-group": "awslogs-${environment}-tis-template",
+          "awslogs-group": "awslogs-${environment}-tis-trainee-notifications",
           "awslogs-region": "eu-west-2",
-          "awslogs-stream-prefix": "awslogs-tis-template"
+          "awslogs-stream-prefix": "awslogs-tis-trainee-notifications"
         }
       },
       "portMappings": [
         {
-          "containerPort": 8080
+          "containerPort": 8211
         }
       ],
       "environment": [
         {
           "name": "TITLE",
-          "value": "tis-template"
+          "value": "tis-trainee-notifications"
         },
         {
           "name": "AWS_REGION",
           "value": "eu-west-2"
         },
         {
-          "name": "SENTRY_ENVIRONMENT",
+          "name": "ENVIRONMENT",
           "value": "${environment}"
         }
       ]
     }
   ],
-  "family": "tis-template-${environment}",
+  "family": "tis-trainee-notifications-${environment}",
   "requiresCompatibilities": [
     "FARGATE"
   ],
