+++ .github/dependabot.yml
@@ -5,8 +5,6 @@ updates:
     schedule:
       interval: weekly
     reviewers:
-      - Health-Education-England/external-admin-backend
-      - Health-Education-England/internal-admin-backend
       - Health-Education-England/trainee-backend
 
   - package-ecosystem: github-actions
@@ -14,9 +12,5 @@ updates:
     schedule:
       interval: weekly
     reviewers:
-      - Health-Education-England/external-admin-backend
-      - Health-Education-England/internal-admin-backend
       - Health-Education-England/trainee-backend
-      - Health-Education-England/external-admin-devops
-      - Health-Education-England/internal-admin-devops
       - Health-Education-England/trainee-devops
