+++ .github/workflows/ci-cd.yml
@@ -6,20 +6,10 @@ on:
       - main
 
 jobs:
-  # TODO: delete this build job and uncomment the ci-cd job once ready to deploy.
-  build:
-    uses: health-education-england/.github/.github/workflows/build-gradle.yml@main
+  ci-cd:
+    name: Build and deploy
+    uses: health-education-england/.github/.github/workflows/ci-cd-gradle.yml@main
     with:
-      # TODO: delete or set to true depending on codeartifict requirements
-      use-codeartifact: false
+      cluster-prefix: trainee
     secrets:
       sonar-token: ${{ secrets.SONAR_TOKEN }}
-
-#  ci-cd:
-#    name: Build and deploy
-#    uses: health-education-england/.github/.github/workflows/ci-cd-gradle.yml@main
-#    with:
-#      cluster-prefix: # TODO: add ECS cluster prefix
-#      use-codeartifact: false # TODO: delete or set to true depending on codeartifict requirements
-#    secrets:
-#      sonar-token: ${{ secrets.SONAR_TOKEN }}
